NEULY-119 Changed 'company_valuations.amount' field type.

+++ database/migrations/2020_10_16_090057_create_company_valuations_table.php
@@ -18,7 +18,7 @@ class CreateCompanyValuationsTable extends Migration
             $table->unsignedBigInteger('company_id');
             $table->unsignedBigInteger('acquirer_id')->nullable();
             $table->date('date');
-            $table->decimal('amount');
+            $table->bigInteger('amount');
             $table->text('notes')->nullable();
             $table->timestamps();
